+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * :sparkles: Adds BundleBuilder
 * :fire: Fragmentx instance rename to instantiate, instanceOrigin rename to instantiateOrigin
 * :white_check_mark: Perfect Fragmentx testing
+* :fire: Serevicex tryStart rename to startIfNoRunning, tryStop rename to stopIfRunning
 
 ## v0.6.0